assume integer ID from db as i32 rather than u16

+++ crates/flagrant-api/src/handlers/environments.rs
@@ -15,7 +15,7 @@ pub struct EnvQueryParams {
 
 pub async fn create(
     DbConnection(mut conn): DbConnection,
-    Path(project_id): Path<u16>,
+    Path(project_id): Path<i32>,
     Json(payload): Json<EnvRequestPayload>,
 ) -> Result<Json<Environment>, ServiceError> {
     let project = project::get_by_id(&mut conn, project_id).await?;
@@ -26,7 +26,7 @@ pub async fn create(
 
 pub async fn fetch_by_id(
     DbConnection(mut conn): DbConnection,
-    Path((_project_id, env_id)): Path<(u16, u16)>,
+    Path((_project_id, env_id)): Path<(i32, i32)>,
 ) -> Result<Json<Environment>, ServiceError> {
     let env = environment::get_by_id(&mut conn, env_id).await?;
 
@@ -35,7 +35,7 @@ pub async fn fetch_by_id(
 
 pub async fn fetch_by_name(
     DbConnection(mut conn): DbConnection,
-    Path((project_id, env_name)): Path<(u16, String)>,
+    Path((project_id, env_name)): Path<(i32, String)>,
 ) -> Result<Json<Environment>, ServiceError> {
     let project = project::get_by_id(&mut conn, project_id).await?;
     let env = environment::get_by_name(&mut conn, &project, env_name).await?;
@@ -46,7 +46,7 @@ pub async fn fetch_by_name(
 pub async fn list(
     DbConnection(mut conn): DbConnection,
     Query(params): Query<EnvQueryParams>,
-    Path(project_id): Path<u16>,
+    Path(project_id): Path<i32>,
 ) -> Result<Json<Vec<Environment>>, ServiceError> {
     let project = project::get_by_id(&mut conn, project_id).await?;
     let envs = match params.prefix {
