+++ crates/flagrant-api/src/handlers/features.rs
@@ -15,7 +15,7 @@ pub(crate) struct FeatureQueryParams {
 
 pub async fn create(
     DbConnection(mut conn): DbConnection,
-    Path(environment_id): Path<u16>,
+    Path(environment_id): Path<i32>,
     Json(payload): Json<FeatureRequestPayload>,
 ) -> Result<Json<Feature>, ServiceError> {
     let env = environment::get_by_id(&mut conn, environment_id).await?;
@@ -33,7 +33,7 @@ pub async fn create(
 
 pub async fn fetch_by_id(
     DbConnection(mut conn): DbConnection,
-    Path((environment_id, feature_id)): Path<(u16, u16)>,
+    Path((environment_id, feature_id)): Path<(i32, i32)>,
 ) -> Result<Json<Feature>, ServiceError> {
     let env = environment::get_by_id(&mut conn, environment_id).await?;
     let feature = feature::get_by_id(&mut conn, &env, feature_id).await?;
@@ -43,7 +43,7 @@ pub async fn fetch_by_id(
 
 pub async fn fetch_by_name(
     DbConnection(mut conn): DbConnection,
-    Path((environment_id, feature_name)): Path<(u16, String)>,
+    Path((environment_id, feature_name)): Path<(i32, String)>,
 ) -> Result<Json<Feature>, ServiceError> {
     let env = environment::get_by_id(&mut conn, environment_id).await?;
     let feature = feature::get_by_name(&mut conn, &env, feature_name).await?;
@@ -53,7 +53,7 @@ pub async fn fetch_by_name(
 
 pub async fn update(
     DbConnection(mut conn): DbConnection,
-    Path((environment_id, feature_id)): Path<(u16, u16)>,
+    Path((environment_id, feature_id)): Path<(i32, i32)>,
     Json(payload): Json<FeatureRequestPayload>,
 ) -> Result<Json<()>, ServiceError> {
     let env = environment::get_by_id(&mut conn, environment_id).await?;
@@ -75,7 +75,7 @@ pub async fn update(
 pub async fn list(
     DbConnection(mut conn): DbConnection,
     Query(params): Query<FeatureQueryParams>,
-    Path(environment_id): Path<u16>,
+    Path(environment_id): Path<i32>,
 ) -> Result<Json<Vec<Feature>>, ServiceError> {
     let env = environment::get_by_id(&mut conn, environment_id).await?;
     let features = match params.prefix {
@@ -88,7 +88,7 @@ pub async fn list(
 
 pub async fn delete(
     DbConnection(mut conn): DbConnection,
-    Path((environment_id, feature_id)): Path<(u16, u16)>,
+    Path((environment_id, feature_id)): Path<(i32, i32)>,
 ) -> Result<Json<()>, ServiceError> {
     let env = environment::get_by_id(&mut conn, environment_id).await?;
     let feature = feature::get_by_id(&mut conn, &env, feature_id).await?;
